backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
motor==3.3.2
pymongo==4.6.0
pydantic==2.5.0
pydantic[email]==2.5.0
transformers==4.36.0
torch>=2.0.0
sentencepiece==0.1.99
protobuf==4.25.1
yfinance==0.2.32
aiohttp==3.9.1
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
